Add scaffolding for anomaly checking
This adds a service that runs all the anomaly checks (`AnomalyChecker`),
as well as sub-services that will do the logic for the individual checks
(`DuplicateTransactionAnomaly`, `MissingDataAnomaly`, and
`UnusualSpendingAnomaly`).

`AnomalyChecker` works by going though each of the defined check
classes by using `klass.constantize` to turn the stringed classes to a
class. We then call `.call(transaction)` for each, and if any of the
sub-services flags an anomaly, we immediately return the result.
Otherwise we return a `{ anomaly: false }` so signify that no anomalies
were found for that transaction.

The specs for `AnomalyChecker` only checks that we call
`.call(transaction)` for each type, handling if it returns an anomaly,
and handling if it does not. We don't really care what the logic is for
each of the sub-services, since we will test them individually as they
are implemented.

The sub-services are just stubs, and will be filled in later.