Remove `anomaly` and `anomaly_reason` on `transactions`
We probably want to let `transactions` have many anomalies. With the
current setup, we can only track one anomalies at a time, so if there
are multiple, after one is cleared, another one will show up.

This commit adds a migration that removes these columns. This is safe to
do because we have not used these columns in any way yet.