Add anomaly model and added it's relation to transactions
Adds a migration to add an `anomalies` table, an Anomaly model, and made
transactions have many anomalies.

The Anomaly model has two attributes. `anomaly_type` would be what kind
it is (like if it's a duplicate, or is missing data, or seems like
unusual spending). The `reason` is a description of the anomaly.

Ideally, the attributes of Anomaly should be set by the service that
generates it.