Implement MissingDataAnomaly and wired it up
This implements an simple MissingDataAnomaly service that will flag if
description, amount, date or category is blank.

In order to support this, we need to make sure the presence validation on
description, amount, date and category is not present in the Transaction
model.

This commit also wires up MissingDataAnomaly into AnomalyChecker. The
specs were updated to actually use the MissingDataAnomaly service to
generate the anomaly.

Transaction model was updated to utilize the AnomalyChecker to run the
anomaly checks.

In theory, the Rails backend should be all wired up to generate missing
data anomalies for transactions.

Also updated the transaction form to no longer make all fields required,
so that we can manually enter in Transactions with missing fields.