Add backend functionality for transaction approval
This commits adds a migration to add approval fields to the transactions
table.

These fields are set the following way:

When we create a new transaction and there are no anomalies, approve can be
true, approved_at should be the same as created_at (or the Time.now of
evaluation), and reviewed_by should be "system".

If there are anomalies, approved will be false, approved_at will be nil, and
reviewed_by will be nil.

If the user approves this, then approved will be true, approved at will be when
it happened, and reviewed_by should be "user".

This also updates the approve action in the TransactionsController. When
this is called, we know that the user wants to accept it as is, so we
save the approval attributes on Transaction to be true, by the user at
that time.

This also needs an update on how we run the anomaly checks. We now only
want to auto approve or reset approvals when the reviewed_by is not the user.